complete modularization. fully decoupled. DRY af
