+++ src/hyprlax_main.c
@@ -34,7 +34,7 @@
 #define MAX_CONFIG_LINE_SIZE 512
 
 /* Forward declarations */
-static int hyprlax_load_layer_textures(hyprlax_context_t *ctx);
+/* moved: hyprlax_load_layer_textures in core/render_core.c */
 /* Forward declarations for handlers used before definition */
 void hyprlax_handle_workspace_change_2d(hyprlax_context_t *ctx,
                                        int from_x, int from_y,
@@ -42,11 +42,7 @@ void hyprlax_handle_workspace_change_2d(hyprlax_context_t *ctx,
 /* Helper prototype: parse either --opt=value or --opt value */
 static inline const char* arg_get_val_local(const char *arg, const char *next);
 
-/* Linux event/timer helpers */
-static int create_timerfd_monotonic(void) {
-    int fd = timerfd_create(CLOCK_MONOTONIC, TFD_NONBLOCK | TFD_CLOEXEC);
-    return fd;
-}
+/* Linux event/timer helpers moved to core/event_loop.c */
 
 /* --- Local helpers (refactored from nested functions) --- */
 static bool parse_bool_local(const char *v) {
@@ -98,234 +94,28 @@ static const char* fit_to_string_local(int m) {
     }
 }
 
-static void disarm_timerfd(int fd) {
-    if (fd < 0) return;
-    struct itimerspec its = {0};
-    timerfd_settime(fd, 0, &its, NULL);
-}
+/* disarm_timerfd moved */
 
-static void arm_timerfd_ms(int fd, int initial_ms, int interval_ms) {
-    if (fd < 0) return;
-    struct itimerspec its;
-    its.it_value.tv_sec = initial_ms / 1000;
-    its.it_value.tv_nsec = (initial_ms % 1000) * 1000000L;
-    its.it_interval.tv_sec = interval_ms / 1000;
-    its.it_interval.tv_nsec = (interval_ms % 1000) * 1000000L;
-    timerfd_settime(fd, 0, &its, NULL);
-}
+/* arm_timerfd_ms moved */
 
-static int epoll_add_fd(int epfd, int fd, uint32_t events) {
-    if (epfd < 0 || fd < 0) return -1;
-    struct epoll_event ev = {0};
-    ev.events = events;
-    ev.data.fd = fd;
-    return epoll_ctl(epfd, EPOLL_CTL_ADD, fd, &ev);
-}
+/* epoll_add_fd moved */
 
-static void hyprlax_setup_epoll(hyprlax_context_t *ctx) {
-    if (!ctx) return;
-    ctx->epoll_fd = epoll_create1(EPOLL_CLOEXEC);
-    if (ctx->epoll_fd < 0) return;
+/* hyprlax_setup_epoll moved */
 
-    /* Cache event fds */
-    ctx->platform_event_fd = (ctx->platform && ctx->platform->ops->get_event_fd)
-        ? ctx->platform->ops->get_event_fd() : -1;
-    ctx->compositor_event_fd = (ctx->compositor && ctx->compositor->ops->get_event_fd)
-        ? ctx->compositor->ops->get_event_fd() : -1;
-    ctx->ipc_event_fd = (ctx->ipc_ctx) ? ((ipc_context_t*)ctx->ipc_ctx)->socket_fd : -1;
+/* hyprlax_arm_frame_timer moved */
 
-    /* Create timerfds */
-    ctx->frame_timer_fd = create_timerfd_monotonic();
-    ctx->debounce_timer_fd = create_timerfd_monotonic();
-    ctx->frame_timer_armed = false;
-    ctx->debounce_pending = false;
-    disarm_timerfd(ctx->frame_timer_fd);
-    disarm_timerfd(ctx->debounce_timer_fd);
+/* hyprlax_disarm_frame_timer moved */
 
-    /* Register sources */
-    epoll_add_fd(ctx->epoll_fd, ctx->platform_event_fd, EPOLLIN);
-    epoll_add_fd(ctx->epoll_fd, ctx->compositor_event_fd, EPOLLIN);
-    if (ctx->cursor_event_fd >= 0) {
-        epoll_add_fd(ctx->epoll_fd, ctx->cursor_event_fd, EPOLLIN);
-    }
-    epoll_add_fd(ctx->epoll_fd, ctx->ipc_event_fd, EPOLLIN);
-    epoll_add_fd(ctx->epoll_fd, ctx->frame_timer_fd, EPOLLIN);
-    epoll_add_fd(ctx->epoll_fd, ctx->debounce_timer_fd, EPOLLIN);
-}
+/* hyprlax_arm_debounce moved */
 
-static void hyprlax_arm_frame_timer(hyprlax_context_t *ctx, int fps) {
-    if (!ctx) return;
-    if (fps <= 0) fps = 60;
-    int interval_ms = (int)(1000.0 / (double)fps);
-    if (ctx->frame_timer_fd >= 0) {
-        arm_timerfd_ms(ctx->frame_timer_fd, interval_ms, interval_ms);
-        ctx->frame_timer_armed = true;
-    }
-}
+/* hyprlax_clear_timerfd moved */
 
-static void hyprlax_disarm_frame_timer(hyprlax_context_t *ctx) {
-    if (!ctx) return;
-    if (ctx->frame_timer_fd >= 0) {
-        disarm_timerfd(ctx->frame_timer_fd);
-    }
-    ctx->frame_timer_armed = false;
-}
+/* cursor smoothing moved to core/cursor.c */
 
-static void hyprlax_arm_debounce(hyprlax_context_t *ctx, int debounce_ms) {
-    if (!ctx) return;
-    if (ctx->debounce_timer_fd >= 0) {
-        arm_timerfd_ms(ctx->debounce_timer_fd, debounce_ms, 0);
-        ctx->debounce_pending = true;
-    }
-}
-
-static void hyprlax_clear_timerfd(int fd) {
-    if (fd < 0) return;
-    uint64_t expirations;
-    ssize_t r = read(fd, &expirations, sizeof(expirations));
-    (void)r;
-}
-
-/* Update cursor smoothed normalized values */
-static void cursor_apply_sample(hyprlax_context_t *ctx, float norm_x, float norm_y) {
-    if (!ctx) return;
-    /* Sensitivity */
-    float sx = norm_x * ctx->config.cursor_sensitivity_x;
-    float sy = norm_y * ctx->config.cursor_sensitivity_y;
-
-    /* EMA smoothing */
-    float a = ctx->config.cursor_ema_alpha;
-    if (a < 0.0f) a = 0.0f;
-    if (a > 1.0f) a = 1.0f;
-    ctx->cursor_ema_x = ctx->cursor_ema_x + a * (sx - ctx->cursor_ema_x);
-    ctx->cursor_ema_y = ctx->cursor_ema_y + a * (sy - ctx->cursor_ema_y);
-
-    /* Clamp */
-    if (ctx->cursor_ema_x < -1.0f) ctx->cursor_ema_x = -1.0f;
-    if (ctx->cursor_ema_x >  1.0f) ctx->cursor_ema_x =  1.0f;
-    if (ctx->cursor_ema_y < -1.0f) ctx->cursor_ema_y = -1.0f;
-    if (ctx->cursor_ema_y >  1.0f) ctx->cursor_ema_y =  1.0f;
-
-    ctx->cursor_norm_x = ctx->cursor_ema_x;
-    ctx->cursor_norm_y = ctx->cursor_ema_y;
-}
-
-/* Placeholder: read cursor provider event, if implemented */
-static bool process_cursor_event(hyprlax_context_t *ctx) {
-    if (!ctx) return false;
-    if (ctx->cursor_event_fd >= 0) {
-        hyprlax_clear_timerfd(ctx->cursor_event_fd);
-    }
-
-    double x = 0.0, y = 0.0;
-    bool got_pos = false;
-    if (ctx->config.cursor_follow_global && ctx->compositor && ctx->compositor->ops && ctx->compositor->ops->get_cursor_position) {
-        double cx = 0.0, cy = 0.0;
-        if (ctx->compositor->ops->get_cursor_position(&cx, &cy) == HYPRLAX_SUCCESS) {
-            x = cx; y = cy; got_pos = true;
-            LOG_TRACE("Compositor cursor: x=%.1f, y=%.1f", x, y);
-        }
-    }
-    if (!got_pos && ctx->platform && ctx->platform->ops && ctx->platform->ops->get_cursor_global) {
-        double px = 0.0, py = 0.0;
-        if (ctx->platform->ops->get_cursor_global(&px, &py)) {
-            x = px; y = py; got_pos = true;
-            LOG_TRACE("Platform pointer: x=%.1f, y=%.1f", x, y);
-        }
-    }
-    if (!got_pos) return false;
-
-    int mon_x = 0, mon_y = 0, mon_w = 1920, mon_h = 1080;
-    if (ctx->monitors && ctx->monitors->head) {
-        monitor_instance_t *m = ctx->monitors->head;
-        monitor_instance_t *found = NULL;
-        while (m) {
-            if (x >= m->global_x && x < (m->global_x + m->width) &&
-                y >= m->global_y && y < (m->global_y + m->height)) {
-                found = m; break;
-            }
-            m = m->next;
-        }
-        if (!found) found = ctx->monitors->primary ? ctx->monitors->primary : ctx->monitors->head;
-        if (found) {
-            mon_x = found->global_x; mon_y = found->global_y; mon_w = found->width; mon_h = found->height;
-        }
-    }
-
-    double cx = mon_x + mon_w * 0.5;
-    double cy = mon_y + mon_h * 0.5;
-    double dx = x - cx;
-    double dy = y - cy;
-
-    double dz = ctx->config.cursor_deadzone_px;
-    if (fabs(dx) < dz) dx = 0.0;
-    if (fabs(dy) < dz) dy = 0.0;
-
-    float nx = 0.0f, ny = 0.0f;
-    if (mon_w > 0) nx = (float)(dx / (mon_w * 0.5));
-    if (mon_h > 0) ny = (float)(dy / (mon_h * 0.5));
-    if (nx < -1.0f) nx = -1.0f;
-    if (nx > 1.0f) nx = 1.0f;
-    if (ny < -1.0f) ny = -1.0f;
-    if (ny > 1.0f) ny = 1.0f;
-
-    float prev_x = ctx->cursor_norm_x;
-    float prev_y = ctx->cursor_norm_y;
-    LOG_TRACE("Cursor normalize: mon=(%d,%d %dx%d) center=(%.1f,%.1f) dx=%.1f dy=%.1f -> nx=%.4f ny=%.4f",
-              mon_x, mon_y, mon_w, mon_h, cx, cy, dx, dy, nx, ny);
-
-    cursor_apply_sample(ctx, nx, ny);
-
-    /* Start or update easing animations toward the new smoothed target, if enabled */
-    if (ctx->config.cursor_anim_duration > 0.0) {
-        /* Initialize eased values on first sample to avoid jump-to-zero */
-        if (!ctx->cursor_ease_initialized) {
-            ctx->cursor_eased_x = ctx->cursor_norm_x;
-            ctx->cursor_eased_y = ctx->cursor_norm_y;
-            ctx->cursor_ease_initialized = true;
-        }
-        const float thr = 0.0003f;
-        /* X axis */
-        if (animation_is_active(&ctx->cursor_anim_x)) {
-            /* Do not restart; just retarget if moved enough */
-            if (fabsf(ctx->cursor_anim_x.to_value - ctx->cursor_norm_x) > thr) {
-                ctx->cursor_anim_x.to_value = ctx->cursor_norm_x;
-                ctx->cursor_anim_x.duration = ctx->config.cursor_anim_duration;
-                ctx->cursor_anim_x.easing = ctx->config.cursor_easing;
-            }
-        } else {
-            float ex = ctx->cursor_eased_x;
-            if (fabsf(ctx->cursor_norm_x - ex) > thr) {
-                animation_start(&ctx->cursor_anim_x, ex, ctx->cursor_norm_x,
-                                ctx->config.cursor_anim_duration, ctx->config.cursor_easing);
-            }
-        }
-        /* Y axis */
-        if (animation_is_active(&ctx->cursor_anim_y)) {
-            if (fabsf(ctx->cursor_anim_y.to_value - ctx->cursor_norm_y) > thr) {
-                ctx->cursor_anim_y.to_value = ctx->cursor_norm_y;
-                ctx->cursor_anim_y.duration = ctx->config.cursor_anim_duration;
-                ctx->cursor_anim_y.easing = ctx->config.cursor_easing;
-            }
-        } else {
-            float ey = ctx->cursor_eased_y;
-            if (fabsf(ctx->cursor_norm_y - ey) > thr) {
-                animation_start(&ctx->cursor_anim_y, ey, ctx->cursor_norm_y,
-                                ctx->config.cursor_anim_duration, ctx->config.cursor_easing);
-            }
-        }
-    }
-
-    float dxn = fabsf(ctx->cursor_norm_x - prev_x);
-    float dyn = fabsf(ctx->cursor_norm_y - prev_y);
-
-    /* In debug builds, always schedule a render when we have a position. */
-    if (ctx->config.debug) return true;
-    return (dxn > 0.0015f || dyn > 0.0015f);
-}
+/* cursor events handled in core/event_loop.c via hyprlax_cursor_tick */
 
 /* Apply a compositor workspace event (shared by immediate and debounced paths) */
-static void process_workspace_event(hyprlax_context_t *ctx, const compositor_event_t *comp_event) {
+void process_workspace_event(hyprlax_context_t *ctx, const compositor_event_t *comp_event) {
     if (!ctx || !comp_event) return;
     /* In cursor-only mode, monitors are realized via Wayland output events.
        Skip workspace event processing entirely to avoid any parallax updates. */
@@ -456,129 +246,7 @@ void hyprlax_destroy(hyprlax_context_t *ctx) {
     free(ctx);
 }
 
-/* Resolve path relative to config file directory */
-static char* resolve_config_relative_path(const char *config_path, const char *relative_path) {
-    if (!relative_path) return NULL;
-
-    /* If path is absolute, return a copy */
-    if (relative_path[0] == '/') {
-        return strdup(relative_path);
-    }
-
-    /* Get directory of config file */
-    char *config_copy = strdup(config_path);
-    if (!config_copy) return NULL;
-
-    char *config_dir = dirname(config_copy);
-
-    /* Construct full path */
-    char full_path[PATH_MAX];
-    snprintf( full_path, sizeof(full_path), "%s/%s", config_dir, relative_path);
-
-    free(config_copy);
-
-    /* Resolve to absolute path */
-    char *resolved = realpath(full_path, NULL);
-    if (!resolved) {
-        /* If realpath fails, just return the constructed path */
-        return strdup(full_path);
-    }
-
-    return resolved;
-}
-
-/* Parse config file */
-static int parse_config_file(hyprlax_context_t *ctx, const char *filename) {
-    if (!ctx || !filename) return -1;
-
-    FILE *file = fopen(filename, "r");
-    if (!file) {
-        LOG_ERROR("Cannot open config file: %s", filename);
-        return -1;
-    }
-
-    char line[MAX_CONFIG_LINE_SIZE];
-    int line_num = 0;
-
-    while (fgets(line, MAX_CONFIG_LINE_SIZE, file)) {
-        line_num++;
-
-        /* Skip comments and empty lines */
-        if (line[0] == '#' || line[0] == '\n' || line[0] == '\r') continue;
-
-        /* Remove newline */
-        char *newline = strchr(line, '\n');
-        if (newline) *newline = '\0';
-
-        /* Parse tokens */
-        char *cmd = strtok(line, " \t");
-        if (!cmd) continue;
-
-        if (strcmp(cmd, "layer") == 0) {
-            char *image = strtok(NULL, " \t");
-            char *shift_str = strtok(NULL, " \t");
-            char *opacity_str = strtok(NULL, " \t");
-            char *blur_str = strtok(NULL, " \t");
-
-            if (!image) {
-                LOG_ERROR("Config line %d: layer requires image path", line_num);
-                continue;
-            }
-
-            /* Resolve path relative to config file if needed */
-            char *resolved_image_path = resolve_config_relative_path(filename, image);
-            if (!resolved_image_path) {
-                LOG_ERROR("Failed to resolve image path at line %d: %s", line_num, image);
-                continue;
-            }
-
-            float shift = shift_str ? atof(shift_str) : 1.0f;
-            float opacity = opacity_str ? atof(opacity_str) : 1.0f;
-            float blur = blur_str ? atof(blur_str) : 0.0f;
-
-            LOG_DEBUG("Config parse layer: image=%s, shift=%.2f, opacity=%.2f, blur=%.2f",
-                    resolved_image_path, shift, opacity, blur);
-
-            /* Add the layer */
-            hyprlax_add_layer(ctx, resolved_image_path, shift, opacity, blur);
-            free(resolved_image_path);
-
-        } else if (strcmp(cmd, "duration") == 0) {
-            char *val = strtok(NULL, " \t");
-            if (val) ctx->config.animation_duration = atof(val);
-        } else if (strcmp(cmd, "shift") == 0) {
-            char *val = strtok(NULL, " \t");
-            if (val) ctx->config.shift_pixels = atof(val);
-        } else if (strcmp(cmd, "scale") == 0) {
-            char *val = strtok(NULL, " \t");
-            if (val) ctx->config.scale_factor = atof(val);
-        } else if (strcmp(cmd, "fps") == 0) {
-            char *val = strtok(NULL, " \t");
-            if (val) ctx->config.target_fps = atoi(val);
-        } else if (strcmp(cmd, "easing") == 0) {
-            char *val = strtok(NULL, " \t");
-            if (val) {
-                ctx->config.default_easing = easing_from_string(val);
-            }
-        } else if (strcmp(cmd, "vsync") == 0) {
-            char *val = strtok(NULL, " \t");
-            if (val) {
-                ctx->config.vsync = (atoi(val) != 0);
-            }
-        } else if (strcmp(cmd, "idle_poll_rate") == 0) {
-            char *val = strtok(NULL, " \t");
-            if (val) {
-                ctx->config.idle_poll_rate = atof(val);
-                if (ctx->config.idle_poll_rate < 0.1f || ctx->config.idle_poll_rate > 10.0f) {
-                    ctx->config.idle_poll_rate = 2.0f;
-                }
-            }
-        }
-    }
-
-    fclose(file);
-    return 0;
-}
+/* Legacy config parser removed (TOML-only runtime) */
 
 /* Reload configuration (TOML only) and re-apply layers */
 int hyprlax_reload_config(hyprlax_context_t *ctx) {
@@ -622,7 +290,8 @@ static int parse_arguments(hyprlax_context_t *ctx, int argc, char **argv) {
     const char *verb_env = getenv("HYPRLAX_VERBOSE"); /* optional numeric override 0..4 */
     if (verb_env && *verb_env) {
         int v = atoi(verb_env);
-        if (v < 0) v = 0; if (v > 4) v = 4;
+        if (v < 0) v = 0;
+        if (v > 4) v = 4;
         ctx->config.log_level = v;
         if (v >= 3) ctx->config.debug = true; /* map DEBUG/TRACE to debug behavior */
     } else if (trace_env && *trace_env && strcmp(trace_env, "0") != 0 && strcasecmp(trace_env, "false") != 0) {
@@ -926,7 +595,8 @@ static int parse_arguments(hyprlax_context_t *ctx, int argc, char **argv) {
                             float ts = 1.0f;
                             if (tint_strength_str && *tint_strength_str) {
                                 ts = (float)atof(tint_strength_str);
-                                if (ts < 0.0f) ts = 0.0f; if (ts > 1.0f) ts = 1.0f;
+                                if (ts < 0.0f) ts = 0.0f;
+                                if (ts > 1.0f) ts = 1.0f;
                             }
                             last->tint_strength = ts;
                         }
@@ -1395,35 +1065,7 @@ int hyprlax_init(hyprlax_context_t *ctx, int argc, char **argv) {
     return HYPRLAX_SUCCESS;
 }
 
-/* Load texture from file */
-static inline int is_pow2(int v) { return v > 0 && (v & (v - 1)) == 0; }
-static GLuint load_texture(const char *path, int *width, int *height) {
-    int channels;
-    unsigned char *data = stbi_load(path, width, height, &channels, 4);
-    if (!data) {
-        LOG_ERROR("Failed to load image '%s': %s", path, stbi_failure_reason());
-        return 0;
-    }
-
-    GLuint texture;
-    glGenTextures(1, &texture);
-    glBindTexture(GL_TEXTURE_2D, texture);
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA, *width, *height, 0, GL_RGBA, GL_UNSIGNED_BYTE, data);
-
-    /* NPOT safety for GLES2: only generate mipmaps if both dimensions are powers of two */
-    if (is_pow2(*width) && is_pow2(*height)) {
-        glGenerateMipmap(GL_TEXTURE_2D);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR_MIPMAP_LINEAR);
-    } else {
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
-    }
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
-
-    stbi_image_free(data);
-    return texture;
-}
+/* Load texture moved to core/render_core.c */
 
 /* Add a layer */
 int hyprlax_add_layer(hyprlax_context_t *ctx, const char *image_path,
@@ -1483,40 +1125,7 @@ void hyprlax_remove_layer(hyprlax_context_t *ctx, uint32_t layer_id) {
     ctx->layer_count = layer_list_count(ctx->layers);
 }
 
-/* Load textures for all layers (called after GL init) */
-static int hyprlax_load_layer_textures(hyprlax_context_t *ctx) {
-    if (!ctx) return HYPRLAX_ERROR_INVALID_ARGS;
-
-    int loaded = 0;
-    parallax_layer_t *layer = ctx->layers;
-    while (layer) {
-        if (layer->texture_id == 0 && layer->image_path) {
-            int img_width, img_height;
-            GLuint texture = load_texture(layer->image_path, &img_width, &img_height);
-            if (texture != 0) {
-                layer->texture_id = texture;
-                layer->width = img_width;
-                layer->height = img_height;
-                layer->texture_width = img_width;
-                layer->texture_height = img_height;
-                loaded++;
-                if (ctx->config.debug) {
-                    LOG_DEBUG("Loaded texture for layer: %s (%dx%d)",
-                            layer->image_path, img_width, img_height);
-                }
-            } else {
-                LOG_ERROR("Failed to load texture for layer: %s", layer->image_path);
-            }
-        }
-        layer = layer->next;
-    }
-
-    if (ctx->config.debug && loaded > 0) {
-        LOG_INFO("Loaded %d layer textures", loaded);
-    }
-
-    return HYPRLAX_SUCCESS;
-}
+/* moved to core/render_core.c: hyprlax_load_layer_textures */
 
 /* Handle per-monitor workspace change */
 void hyprlax_handle_monitor_workspace_change(hyprlax_context_t *ctx,
@@ -1622,24 +1231,7 @@ void hyprlax_handle_workspace_change_2d(hyprlax_context_t *ctx,
     ctx->workspace_offset_y = target_y;
 }
 
-/* Get current time in seconds */
-static double get_time(void) {
-    struct timespec ts;
-    clock_gettime(CLOCK_MONOTONIC, &ts);
-    return ts.tv_sec + ts.tv_nsec / 1000000000.0;
-}
-
-/* Calculate optimal scale factor for parallax effect */
-static float calculate_scale_factor(float shift_pixels, int max_workspaces, int viewport_width) {
-    /* Scale factor determines how much larger the image is than the viewport
-     * We need: image_width = viewport_width * scale_factor
-     * And: (scale_factor - 1) * viewport_width >= total_shift_needed */
-    float total_shift_needed = (max_workspaces - 1) * shift_pixels;
-    float min_scale_factor = 1.0f + (total_shift_needed / (float)viewport_width);
-
-    /* Return at least 1.5x for a nice parallax effect */
-    return fmaxf(min_scale_factor, 1.5f);
-}
+/* time helpers moved; scale computations handled by renderer */
 
 /* Update layers */
 void hyprlax_update_layers(hyprlax_context_t *ctx, double current_time) {
@@ -1652,620 +1244,11 @@ void hyprlax_update_layers(hyprlax_context_t *ctx, double current_time) {
     }
 }
 
-/* Render frame to a specific monitor */
-static void hyprlax_render_monitor(hyprlax_context_t *ctx, monitor_instance_t *monitor) {
-    if (!ctx || !ctx->renderer || !monitor) {
-        LOG_TRACE("Skipping render: ctx=%p, renderer=%p, monitor=%p", ctx, ctx ? ctx->renderer : NULL, monitor);
-        return;
-    }
+/* hyprlax_render_frame moved to core/render_core.c */
 
-    /* Skip if monitor doesn't have an EGL surface */
-    if (!monitor->egl_surface) {
-        LOG_WARN("Monitor %s has no EGL surface", monitor->name);
-        return;
-    }
+/* has_active_animations removed (handled in core/event_loop.c) */
 
-    /* Monitor rendering debug - commented out for performance
-    LOG_TRACE("Rendering to monitor %s (%dx%d, surface=%p)",
-              monitor->name, monitor->width, monitor->height, monitor->egl_surface); */
-
-    /* Switch to this monitor's EGL surface */
-    if (gles2_make_current(monitor->egl_surface) != HYPRLAX_SUCCESS) {
-        LOG_ERROR("Failed to make EGL surface current for monitor %s", monitor->name);
-        return;
-    }
-
-    /* Set viewport for this monitor */
-    glViewport(0, 0, monitor->width * monitor->scale, monitor->height * monitor->scale);
-    /* Viewport debug - commented out for performance
-    LOG_TRACE("Set viewport: %dx%d (scale=%d)",
-              monitor->width * monitor->scale, monitor->height * monitor->scale, monitor->scale); */
-
-    /* Optional profiling */
-    static int s_profile = -1;
-    if (s_profile == -1) {
-        const char *p = getenv("HYPRLAX_PROFILE");
-        s_profile = (p && *p) ? 1 : 0;
-    }
-    double t_draw_start = 0.0;
-    double t_present_start = 0.0;
-
-    /* Clear and prepare for rendering */
-    RENDERER_BEGIN_FRAME(ctx->renderer);
-    RENDERER_CLEAR(ctx->renderer, 0.0f, 0.0f, 0.0f, 1.0f);
-
-    /* Enable/disable blending only when needed to avoid redundant state changes */
-    static bool s_blend_enabled = false;
-    bool need_blend = (ctx->layer_count > 1);
-    if (need_blend != s_blend_enabled) {
-        if (need_blend) {
-            glEnable(GL_BLEND);
-            glBlendFunc(GL_ONE, GL_ONE_MINUS_SRC_ALPHA);
-        } else {
-            glDisable(GL_BLEND);
-        }
-        s_blend_enabled = need_blend;
-    }
-
-    /* Calculate scale factor for this monitor */
-    float scale_factor = ctx->config.scale_factor;
-    if (scale_factor <= 1.0f) {
-        scale_factor = calculate_scale_factor(
-            ctx->config.shift_pixels, 10, monitor->width);
-    }
-
-    /* Render each layer with this monitor's offset */
-    if (s_profile) t_draw_start = get_time();
-    parallax_layer_t *layer = ctx->layers;
-    int layer_num = 0;
-    /* Layer count debug - commented out for performance
-    LOG_TRACE("Rendering %d layers for monitor %s", ctx->layer_count, monitor->name); */
-    while (layer) {
-        /* Attempt lazy texture load once we hit the layer in draw path */
-        if (layer->texture_id == 0 && layer->image_path && ctx->renderer && ctx->renderer->initialized) {
-            int lw = 0, lh = 0;
-            GLuint tid = load_texture(layer->image_path, &lw, &lh);
-            if (tid != 0) {
-                layer->texture_id = tid;
-                layer->width = lw;
-                layer->height = lh;
-                layer->texture_width = lw;
-                layer->texture_height = lh;
-                LOG_DEBUG("[LAZY] Loaded texture for layer %u (%s): %dx%d tex=%u",
-                          layer->id, layer->image_path, lw, lh, tid);
-            } else {
-                LOG_WARN("[LAZY] Failed to load texture for layer %u: %s", layer->id, layer->image_path);
-                ctx->deferred_render_needed = true; /* request another frame to retry */
-            }
-        }
-
-        if (layer->hidden || layer->texture_id == 0) {
-            layer = layer->next;
-            continue;
-        }
-
-        /* Base offsets from workspace animation */
-        float workspace_x = layer->offset_x;
-        float workspace_y = layer->offset_y;
-
-        /* Apply optional workspace inversion (global xor layer) */
-        if (ctx->config.invert_workspace_x ^ layer->invert_workspace_x) {
-            workspace_x = -workspace_x;
-        }
-        if (ctx->config.invert_workspace_y ^ layer->invert_workspace_y) {
-            workspace_y = -workspace_y;
-        }
-
-        /* Cursor-driven offsets (normalized -> pixels) */
-        float cursor_x_px = ctx->cursor_eased_x * ctx->config.parallax_max_offset_x * layer->shift_multiplier_x;
-        float cursor_y_px = ctx->cursor_eased_y * ctx->config.parallax_max_offset_y * layer->shift_multiplier_y;
-
-        if (ctx->config.invert_cursor_x ^ layer->invert_cursor_x) {
-            cursor_x_px = -cursor_x_px;
-        }
-        if (ctx->config.invert_cursor_y ^ layer->invert_cursor_y) {
-            cursor_y_px = -cursor_y_px;
-        }
-
-        /* Blend according to selected mode */
-        float offset_x = 0.0f;
-        float offset_y = 0.0f;
-        switch (ctx->config.parallax_mode) {
-            case PARALLAX_WORKSPACE:
-                offset_x = workspace_x;
-                offset_y = workspace_y;
-                break;
-            case PARALLAX_CURSOR:
-                offset_x = cursor_x_px * ctx->config.parallax_cursor_weight;
-                offset_y = cursor_y_px * ctx->config.parallax_cursor_weight;
-                break;
-            case PARALLAX_HYBRID:
-            default:
-                offset_x = workspace_x * ctx->config.parallax_workspace_weight +
-                           cursor_x_px * ctx->config.parallax_cursor_weight;
-                offset_y = workspace_y * ctx->config.parallax_workspace_weight +
-                           cursor_y_px * ctx->config.parallax_cursor_weight;
-                break;
-        }
-
-        if (ctx->config.debug) {
-            LOG_TRACE("Layer %u: ws(%.1f,%.1f) cur(%.1f,%.1f) -> off_px(%.1f,%.1f) off_norm(%.5f,%.5f)",
-                      layer->id,
-                      workspace_x, workspace_y,
-                      cursor_x_px, cursor_y_px,
-                      offset_x, offset_y,
-                      offset_x / (float)monitor->width,
-                      offset_y / (float)monitor->height);
-        }
-
-        if (getenv("HYPRLAX_DEBUG")) {
-            fprintf(stderr,
-                    "[DEBUG] layer id=%u ws(%.2f,%.2f) cur(%.2f,%.2f) -> off_px(%.2f,%.2f) off_norm(%.5f,%.5f)\n",
-                    layer->id,
-                    workspace_x, workspace_y,
-                    cursor_x_px, cursor_y_px,
-                    offset_x, offset_y,
-                    offset_x / (float)monitor->width,
-                    offset_y / (float)monitor->height);
-        }
-
-        /* Create texture struct for drawing */
-        texture_t tex = {
-            .id = layer->texture_id,
-            .width = layer->width,
-            .height = layer->height
-        };
-
-        /* Bind texture */
-        if (ctx->renderer->ops->bind_texture) {
-            ctx->renderer->ops->bind_texture(&tex, 0);
-        }
-
-        const char *force_legacy = getenv("HYPRLAX_FORCE_LEGACY");
-        if (ctx->renderer->ops->draw_layer_ex && !(force_legacy && *force_legacy)) {
-            int eff_over = (layer->overflow_mode >= 0) ? layer->overflow_mode : ctx->config.render_overflow_mode;
-            int eff_tile_x;
-            int eff_tile_y;
-            if (layer->tile_x >= 0) eff_tile_x = layer->tile_x;
-            else {
-                if (eff_over == 1 /* repeat */ || eff_over == 2 /* repeat_x */) eff_tile_x = 1;
-                else if (eff_over == 3 /* repeat_y */) eff_tile_x = 0;
-                else eff_tile_x = ctx->config.render_tile_x;
-            }
-            if (layer->tile_y >= 0) eff_tile_y = layer->tile_y;
-            else {
-                if (eff_over == 1 /* repeat */ || eff_over == 3 /* repeat_y */) eff_tile_y = 1;
-                else if (eff_over == 2 /* repeat_x */) eff_tile_y = 0;
-                else eff_tile_y = ctx->config.render_tile_y;
-            }
-            renderer_layer_params_t p = {
-                .fit_mode = layer->fit_mode,
-                .content_scale = layer->content_scale,
-                .align_x = layer->align_x,
-                .align_y = layer->align_y,
-                .base_uv_x = layer->base_uv_x,
-                .base_uv_y = layer->base_uv_y,
-                .overflow_mode = eff_over,
-                .margin_px_x = (layer->margin_px_x != 0.0f || layer->margin_px_y != 0.0f) ? layer->margin_px_x : ctx->config.render_margin_px_x,
-                .margin_px_y = (layer->margin_px_x != 0.0f || layer->margin_px_y != 0.0f) ? layer->margin_px_y : ctx->config.render_margin_px_y,
-                .tile_x = eff_tile_x,
-                .tile_y = eff_tile_y,
-                .auto_safe_norm_x = (ctx->config.parallax_max_offset_x > 0.0f && (eff_tile_x == 0) && (eff_over == 4))
-                    ? (ctx->config.parallax_max_offset_x / (float)monitor->width) : 0.0f,
-                .auto_safe_norm_y = (ctx->config.parallax_max_offset_y > 0.0f && (eff_tile_y == 0) && (eff_over == 4))
-                    ? (ctx->config.parallax_max_offset_y / (float)monitor->height) : 0.0f,
-                .tint_r = layer->tint_r,
-                .tint_g = layer->tint_g,
-                .tint_b = layer->tint_b,
-                .tint_strength = layer->tint_strength,
-            };
-            ctx->renderer->ops->draw_layer_ex(
-                &tex,
-                offset_x / monitor->width,
-                offset_y / monitor->height,
-                layer->opacity,
-                layer->blur_amount,
-                &p
-            );
-        } else if (ctx->renderer->ops->draw_layer) {
-            /* Draw the layer with offset and effects */
-            /* Layer drawing debug - commented out for performance
-            LOG_TRACE("Drawing layer %d (tex_id=%u, offset=%.3f,%.3f, opacity=%.2f)",
-                     layer_num++, layer->texture_id,
-                     offset_x / monitor->width, offset_y / monitor->height, layer->opacity); */
-            layer_num++;
-            ctx->renderer->ops->draw_layer(
-                &tex,
-                offset_x / monitor->width,  /* Normalized X offset */
-                offset_y / monitor->height, /* Normalized Y offset */
-                layer->opacity,
-                layer->blur_amount
-            );
-        }
-
-        layer = layer->next;
-    }
-
-    /* Present this monitor's frame */
-    /* Present debug - commented out for performance
-    LOG_TRACE("Ending frame and presenting for monitor %s", monitor->name); */
-    RENDERER_END_FRAME(ctx->renderer);
-    double t_draw_end = s_profile ? get_time() : 0.0;
-    if (s_profile) t_present_start = t_draw_end;
-    RENDERER_PRESENT(ctx->renderer);
-    double t_present_end = s_profile ? get_time() : 0.0;
-
-    if (s_profile && ctx->config.debug) {
-        double draw_ms = (t_draw_end - t_draw_start) * 1000.0;
-        double present_ms = (t_present_end - t_present_start) * 1000.0;
-        LOG_DEBUG("[PROFILE] monitor=%s draw=%.2f ms present=%.2f ms", monitor->name, draw_ms, present_ms);
-    }
-
-    /* Commit the Wayland surface to make the frame visible */
-    if (monitor->wl_surface && ctx->platform && ctx->platform->ops && ctx->platform->ops->commit_monitor_surface) {
-        ctx->platform->ops->commit_monitor_surface(monitor);
-        /* Commit debug - commented out for performance
-        LOG_TRACE("Committed surface for monitor %s", monitor->name); */
-    }
-
-    /* Frame presented debug - commented out for performance
-    LOG_TRACE("Frame presented for monitor %s", monitor->name); */
-}
-
-/* Render frame to all monitors */
-void hyprlax_render_frame(hyprlax_context_t *ctx) {
-    if (!ctx || !ctx->renderer) {
-        LOG_ERROR("render_frame: No renderer available");
-        return;
-    }
-
-    /* Track if this frame is actually different from last frame (unused placeholder removed) */
-
-    /* Always use monitor list - single monitor is just count=1 */
-    if (!ctx->monitors || ctx->monitors->count == 0) {
-        LOG_WARN("No monitors available for rendering");
-        /* Keep running to allow Wayland to finish output discovery */
-        return;
-    }
-
-    /* Monitor count debug - commented out for performance
-    LOG_TRACE("Rendering frame to %d monitor(s)", ctx->monitors->count); */
-
-    /* Update cursor easing state (if enabled) once per frame */
-    double now_time = get_time();
-    if (ctx->config.cursor_anim_duration > 0.0) {
-        if (animation_is_active(&ctx->cursor_anim_x))
-            ctx->cursor_eased_x = animation_evaluate(&ctx->cursor_anim_x, now_time);
-        else
-            ctx->cursor_eased_x = ctx->cursor_norm_x;
-        if (animation_is_active(&ctx->cursor_anim_y))
-            ctx->cursor_eased_y = animation_evaluate(&ctx->cursor_anim_y, now_time);
-        else
-            ctx->cursor_eased_y = ctx->cursor_norm_y;
-    } else {
-        ctx->cursor_eased_x = ctx->cursor_norm_x;
-        ctx->cursor_eased_y = ctx->cursor_norm_y;
-    }
-
-    /* Render to each monitor */
-    monitor_instance_t *monitor = ctx->monitors->head;
-    while (monitor) {
-        /* Monitor call debug - commented out for performance
-        LOG_TRACE("Calling hyprlax_render_monitor for %s", monitor->name); */
-        hyprlax_render_monitor(ctx, monitor);
-        monitor = monitor->next;
-    }
-
-    /* After presenting a frame, diagnostics no longer snapshot prev offsets
-       to avoid touching core layer structs. */
-}
-
-/* Check if any layer has active animations */
-static bool has_active_animations(hyprlax_context_t *ctx) {
-    if (!ctx) return false;
-
-    /* Check layer animations */
-    parallax_layer_t *layer = ctx->layers;
-    while (layer) {
-        if (animation_is_active(&layer->x_animation) ||
-            animation_is_active(&layer->y_animation)) {
-            return true;
-        }
-        layer = layer->next;
-    }
-
-    /* Check monitor animations */
-    if (ctx->monitors) {
-        monitor_instance_t *monitor = ctx->monitors->head;
-        while (monitor) {
-            if (monitor->animating) {
-                return true;
-            }
-            monitor = monitor->next;
-        }
-    }
-
-    return false;
-}
-
-/* Main run loop */
-int hyprlax_run(hyprlax_context_t *ctx) {
-    if (!ctx) return HYPRLAX_ERROR_INVALID_ARGS;
-
-    if (ctx->config.debug) {
-        LOG_DEBUG("Starting main loop (target FPS: %d)", ctx->config.target_fps);
-    }
-
-    /* Optional render diagnostics (why are we rendering while idle) */
-    static int s_render_diag = -1;
-    if (s_render_diag == -1) {
-        const char *p = getenv("HYPRLAX_RENDER_DIAG");
-        s_render_diag = (p && *p) ? 1 : 0;
-    }
-
-    double last_render_time = get_time();
-    double last_frame_time = last_render_time;
-    double frame_time = 1.0 / (double)(ctx->config.target_fps > 0 ? ctx->config.target_fps : 60);
-    int prev_target_fps = ctx->config.target_fps;
-    int frame_count = 0;
-    double debug_timer = 0.0;
-    bool needs_render = true;  /* Render first frame */
-
-    if (ctx->config.debug) {
-        LOG_DEBUG("Frame time calculated: %.6f seconds (target FPS: %d)",
-                  frame_time, ctx->config.target_fps);
-    }
-
-    while (ctx->running) {
-        /* Recompute frame time from current config to honor live FPS changes */
-        int current_fps = ctx->config.target_fps;
-        if (current_fps <= 0) current_fps = 60;
-        if (current_fps != prev_target_fps) {
-            /* If using timerfd (no frame callbacks), re-arm at new interval */
-            const char *fc_env = getenv("HYPRLAX_FRAME_CALLBACK");
-            bool use_frame_callback = (fc_env && *fc_env);
-            if (!use_frame_callback) {
-                hyprlax_arm_frame_timer(ctx, current_fps);
-            }
-            prev_target_fps = current_fps;
-        }
-        frame_time = 1.0 / (double)current_fps;
-
-        /* Reset per-iteration diagnostic flags */
-        bool diag_resize = false;
-        bool diag_ipc = false;
-        bool diag_comp = false;
-        bool diag_final = false;
-
-        double current_time = get_time();
-        ctx->delta_time = current_time - last_frame_time;
-        last_frame_time = current_time;
-
-        /* Frame timing debug - disabled for performance
-        if (ctx->config.debug && frame_count < 5) {
-            LOG_TRACE("Frame %d: delta=%.4f, frame_time=%.4f",
-                   frame_count, ctx->delta_time, frame_time);
-        } */
-
-        /* Poll platform events */
-        platform_event_t platform_event;
-        /* Platform events debug - disabled for performance
-        if (ctx->config.debug && frame_count < 5) {
-            LOG_TRACE("Polling platform events");
-        } */
-        if (PLATFORM_POLL_EVENTS(ctx->platform, &platform_event) == HYPRLAX_SUCCESS) {
-            switch (platform_event.type) {
-                case PLATFORM_EVENT_CLOSE:
-                    ctx->running = false;
-                    break;
-                case PLATFORM_EVENT_RESIZE:
-                    hyprlax_handle_resize(ctx,
-                                        platform_event.data.resize.width,
-                                        platform_event.data.resize.height);
-                    needs_render = true;  /* Window resize requires re-render */
-                    diag_resize = true;
-                    break;
-                default:
-                    break;
-            }
-        }
-
-        /* Poll IPC commands */
-        if (ctx->ipc_ctx && ipc_process_commands((ipc_context_t*)ctx->ipc_ctx)) {
-            /* IPC processed - layers may have changed */
-            needs_render = true;  /* IPC commands may have changed layers */
-            diag_ipc = true;
-            if (ctx->config.debug) {
-                LOG_TRACE("IPC command processed");
-            }
-        }
-
-        /* Poll compositor events */
-        if (ctx->compositor && ctx->compositor->ops->poll_events) {
-            compositor_event_t comp_event;
-            int poll_result = ctx->compositor->ops->poll_events(&comp_event);
-            if (poll_result == HYPRLAX_SUCCESS) {
-                if (comp_event.type == COMPOSITOR_EVENT_WORKSPACE_CHANGE) {
-                    diag_comp = true;
-                    /* Immediate reaction: apply event and request render */
-                    process_workspace_event(ctx, &comp_event);
-                    needs_render = true;
-                }
-            }
-        }
-
-        /* Check if animations are active */
-        bool animations_active = has_active_animations(ctx);
-
-        /* While animating, always need to render; with frame callbacks, wait for compositor pacing */
-        const char *use_fc = getenv("HYPRLAX_FRAME_CALLBACK");
-        if (animations_active) {
-            if (use_fc && *use_fc && ctx->monitors) {
-                /* Render when at least one monitor's frame callback has fired */
-                bool can_render = false;
-                monitor_instance_t *m = ctx->monitors->head;
-                while (m) {
-                    if (!m->frame_pending) { can_render = true; break; }
-                    m = m->next;
-                }
-                needs_render = needs_render || can_render;
-            } else {
-                needs_render = true;
-            }
-        }
-
-        /* Only update animations if they're active */
-        if (animations_active) {
-            /* Update layer animations */
-            hyprlax_update_layers(ctx, current_time);
-
-            /* Update monitor animations for parallax */
-            if (ctx->monitors) {
-                monitor_instance_t *monitor = ctx->monitors->head;
-                while (monitor) {
-                    monitor_update_animation(monitor, current_time);
-                    monitor = monitor->next;
-                }
-            }
-
-            /* Re-check if animations are still active after update */
-            bool still_animating = has_active_animations(ctx);
-
-            /* Ensure one final frame when animations complete */
-            if (!still_animating && animations_active) {
-                needs_render = true;  /* Final frame */
-                diag_final = true;
-            }
-
-            /* Update the animations_active flag for next iteration */
-            animations_active = still_animating;
-        }
-
-        /* Calculate time since last render */
-        double time_since_render = current_time - last_render_time;
-
-        /* Render frame if:
-         * 1. We need to render (animation active, workspace changed, etc.)
-         * 2. Enough time has passed since last frame (respecting target FPS)
-         */
-        if (needs_render && ((use_fc && *use_fc) ? true : (time_since_render >= frame_time))) {
-            if (ctx->config.debug) {
-                static double last_log_time = 0;
-                if (current_time - last_log_time > 1.0) {
-                    LOG_DEBUG("RENDERING: animations=%d, time_since_render=%.3f",
-                              animations_active, time_since_render);
-                    last_log_time = current_time;
-                }
-            }
-            if (s_render_diag && !animations_active) {
-                /* Log first layer offsets to aid diagnosis without prev snapshot */
-                float first_x = 0.0f, first_y = 0.0f;
-                if (ctx->layers) {
-                    first_x = ctx->layers->offset_x;
-                    first_y = ctx->layers->offset_y;
-                }
-                LOG_DEBUG("[RENDER_DIAG] idle render: resize=%d ipc=%d comp=%d final=%d tsr=%.3f ft=%.3f layers=%d first_offset=%.4f,%.4f",
-                          diag_resize, diag_ipc, diag_comp, diag_final,
-                          time_since_render, frame_time, ctx->layer_count,
-                          first_x, first_y);
-            }
-    hyprlax_render_frame(ctx);
-    ctx->fps = 1.0 / time_since_render;
-    last_render_time = current_time;
-    frame_count++;
-
-    /* Clear needs_render flag; request another frame if deferred work pending */
-    needs_render = ctx->deferred_render_needed;
-    ctx->deferred_render_needed = false;
-
-            /* Print FPS every second in debug mode */
-            if (ctx->config.debug) {
-                debug_timer += time_since_render;
-                if (debug_timer >= 1.0) {
-                    LOG_DEBUG("FPS: %.1f, Layers: %d, Animations: %s",
-                             ctx->fps, ctx->layer_count,
-                             animations_active ? "active" : "idle");
-                    debug_timer = 0.0;
-                }
-            }
-        } else {
-            /* Event-driven wait using epoll/timerfd (fallback to idle sleep) */
-            const char *fc_env = getenv("HYPRLAX_FRAME_CALLBACK");
-            bool use_frame_callback = (fc_env && *fc_env);
-
-            /* Manage frame timer when not using frame callbacks */
-            if (!use_frame_callback) {
-                if (has_active_animations(ctx)) {
-                    if (!ctx->frame_timer_armed) {
-                        hyprlax_arm_frame_timer(ctx, ctx->config.target_fps);
-                    }
-                } else {
-                    /* No animations: disarm periodic timer */
-                    if (ctx->frame_timer_armed) {
-                        hyprlax_disarm_frame_timer(ctx);
-                    }
-                    /* If a frame is needed but time gate not yet reached, arm one-shot to wake */
-                    if (needs_render) {
-                        double target_wake_time = last_render_time + frame_time;
-                        double remain = target_wake_time - current_time;
-                        int remain_ms = (int)(remain * 1000.0);
-                        if (remain_ms < 1) remain_ms = 1;
-                        arm_timerfd_ms(ctx->frame_timer_fd, remain_ms, 0); /* one-shot */
-                    }
-                }
-            }
-
-            if (ctx->epoll_fd >= 0) {
-                struct epoll_event evlist[6];
-                int n = epoll_wait(ctx->epoll_fd, evlist, 6, -1);
-                if (n < 0) {
-                    if (errno == EINTR) {
-                        if (!ctx->running) break; /* interrupted by signal; exit loop */
-                        continue;
-                    }
-                    /* Other epoll error: avoid tight loop */
-                    continue;
-                }
-                if (n > 0) {
-                    for (int i = 0; i < n; i++) {
-                        int fd = evlist[i].data.fd;
-                        if (fd == ctx->debounce_timer_fd) {
-                            hyprlax_clear_timerfd(fd);
-                            if (ctx->debounce_pending) {
-                                ctx->debounce_pending = false;
-                                process_workspace_event(ctx, &ctx->pending_event);
-                                needs_render = true; /* Start animation */
-                            }
-                        } else if (fd == ctx->frame_timer_fd) {
-                            hyprlax_clear_timerfd(fd);
-                            needs_render = true; /* Time to render next frame */
-                        } else if (fd == ctx->cursor_event_fd) {
-                            /* Update cursor state; request a frame if changed */
-                            if (process_cursor_event(ctx)) {
-                                needs_render = true;
-                            }
-                        } else {
-                            /* Wayland/compositor/IPC fds wake main loop; handled next iteration */
-                        }
-                    }
-                }
-                /* Loop to process events immediately in next iteration */
-                continue;
-            } else {
-                /* Fallback: gentle idle sleep */
-                double sleep_time = 1.0 / ctx->config.idle_poll_rate;
-                struct timespec ts;
-                ts.tv_sec = (time_t)(sleep_time);
-                ts.tv_nsec = (long)((sleep_time - (time_t)sleep_time) * 1e9);
-                if (ts.tv_nsec > 999999999) ts.tv_nsec = 999999999;
-                if (ts.tv_nsec < 0) ts.tv_nsec = 0;
-                nanosleep(&ts, NULL);
-            }
-        }
-    }
-
-    return HYPRLAX_SUCCESS;
-}
+/* no-op (run loop lives in core/event_loop.c) */
 
 /* Handle resize */
 void hyprlax_handle_resize(hyprlax_context_t *ctx, int width, int height) {
@@ -2390,7 +1373,9 @@ int hyprlax_runtime_set_property(hyprlax_context_t *ctx, const char *property, c
         if (strcmp(leaf, "align.y") == 0) { layer->align_y = atof(value); if (layer->align_y<0) layer->align_y=0; if (layer->align_y>1) layer->align_y=1; return 0; }
         if (strcmp(leaf, "overflow") == 0) {
             int m = overflow_from_string_local(value);
-            if (m == -2) return -1; layer->overflow_mode = m; return 0;
+            if (m == -2) return -1;
+            layer->overflow_mode = m;
+            return 0;
         }
         if (strcmp(leaf, "tile.x") == 0) { layer->tile_x = parse_bool_local(value) ? 1 : 0; return 0; }
         if (strcmp(leaf, "tile.y") == 0) { layer->tile_y = parse_bool_local(value) ? 1 : 0; return 0; }
@@ -2448,7 +1433,9 @@ int hyprlax_runtime_set_property(hyprlax_context_t *ctx, const char *property, c
     /* Render properties (global) */
     if (strcmp(property, "render.overflow") == 0) {
         int m = overflow_from_string_local(value);
-        if (m == -2) return -1; ctx->config.render_overflow_mode = m; return 0;
+        if (m == -2) return -1;
+        ctx->config.render_overflow_mode = m;
+        return 0;
     }
     if (strcmp(property, "render.tile.x") == 0) { ctx->config.render_tile_x = parse_bool_local(value) ? 1 : 0; return 0; }
     if (strcmp(property, "render.tile.y") == 0) { ctx->config.render_tile_y = parse_bool_local(value) ? 1 : 0; return 0; }
@@ -2463,7 +1450,8 @@ int hyprlax_runtime_get_property(hyprlax_context_t *ctx, const char *property, c
     /* Per-layer get: layer.<id>.* */
     if (strncmp(property, "layer.", 6) == 0) {
         const char *p = property + 6; char *end=NULL; long lid=strtol(p,&end,10);
-        if (lid <= 0 || !end || *end != '.') return -1; const char *leaf = end+1;
+        if (lid <= 0 || !end || *end != '.') return -1;
+        const char *leaf = end+1;
         parallax_layer_t *layer = layer_list_find(ctx->layers, (uint32_t)lid);
         if (!layer) return -1;
         if (strcmp(leaf, "hidden") == 0) { W("%s", layer->hidden?"true":"false"); return 0; }
@@ -2519,17 +1507,7 @@ int hyprlax_runtime_get_property(hyprlax_context_t *ctx, const char *property, c
     #undef W
     return -1;
 }
-#ifndef WEAK_CORE_LAYER_STUBS
-__attribute__((weak)) parallax_layer_t* layer_list_find(parallax_layer_t *head, uint32_t layer_id) {
-    (void)head; (void)layer_id; return NULL;
-}
-__attribute__((weak)) parallax_layer_t* layer_list_remove(parallax_layer_t *head, uint32_t layer_id) {
-    (void)layer_id; return head;
-}
-__attribute__((weak)) int layer_list_count(parallax_layer_t *head) {
-    int c=0; for (; head; head=head->next) c++; return c;
-}
-#endif
+/* layer list stubs removed */
 /* Helper to read either --opt=value or --opt value */
 static inline const char* arg_get_val_local(const char *arg, const char *next) {
     if (!arg) return NULL;
